requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
Pillow>=10.0.0
requests>=2.31.0
plotly>=5.15.0
prophet>=1.1.4
beautifulsoup4>=4.12.0
openai>=1.0.0
httpx>=0.24.0
